Enhance UI Login Page and separate logic to hook
